CINEMATIC/requirements.txt
Django>=4.2.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Production server
gunicorn>=21.2.0

# Environment variables
python-dotenv>=1.0.0

# Optional: PostgreSQL support (uncomment if using PostgreSQL)
# psycopg2-binary>=2.9.0
